CentralizedTelescope: Remove init.rs, moving functions to impl block in Params.rs
